requirements.txt
Package           Version
----------------- -------
blinker           1.9.0
click             8.1.7
Flask             2.3.3
Flask-Cors        4.0.0
Flask-SQLAlchemy  3.1.1
itsdangerous      2.2.0
Jinja2            3.1.4
MarkupSafe        3.0.2
pip               22.3.1
python-dotenv     1.0.0
setuptools        65.5.0
SQLAlchemy        2.0.36
typing_extensions 4.12.2
Werkzeug          3.1.3
